requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.5.0
streamlit>=1.25.0
numpy>=1.24.0 